Refactor launcher livedata to flow and edited the unit test.
