+++ mobile/src/main/java/com/elbehiry/delish/ui/launcher/LauncherActivity.kt
@@ -24,8 +24,8 @@ import com.elbehiry.delish.ui.main.launchMainActivity
 import com.elbehiry.delish.ui.onboarding.launchOnBoardingActivity
 import com.elbehiry.delish.ui.theme.DelishComposeTheme
 import dagger.hilt.android.AndroidEntryPoint
-import com.elbehiry.delish.ui.launcher.LauncherViewModel.LaunchDestination.MAIN_ACTIVITY
-import com.elbehiry.delish.ui.launcher.LauncherViewModel.LaunchDestination.ONBOARDING
+import com.elbehiry.delish.ui.launcher.LaunchDestination.MAIN_ACTIVITY
+import com.elbehiry.delish.ui.launcher.LaunchDestination.ON_BOARDING
 import com.elbehiry.delish.ui.util.checkAllMatched
 
 @AndroidEntryPoint
@@ -41,7 +41,7 @@ class LauncherActivity : ComponentActivity() {
                     onLauncherComplete = { destination ->
                         when (destination) {
                             MAIN_ACTIVITY -> launchMainActivity(context = this)
-                            ONBOARDING -> launchOnBoardingActivity(context = this)
+                            ON_BOARDING -> launchOnBoardingActivity(context = this)
                         }.checkAllMatched
                         finish()
                     }
